{{ cookiecutter.service_name }}/requirements.txt
overrides==6.1.0
paho-mqtt==1.6.1
protobuf==3.19
python-dotenv==0.20.0
typing-utils==0.1.0
pyESDL==22.11.2
influxdb~=5.3.1
numpy==1.20.2